Fix Line Highlight Style
The `.hll` class is missing the leading `.` and a comment explaining the
purpose of the class causing line highlights not to have the desired
appearance.

This commit fixes the class added by pygments.rb to highlight a line
within the `github.css` example docs because the Hologram gem links
directly to the file which most people using Hologram are likely to copy
for their own syntax highlighting purposes.